@@ -2,8 +2,8 @@ package com.doiscs.sisman.domain.model.entity;
 
 import java.util.List;
 
-import javax.persistence.Entity;
-import javax.persistence.OneToMany;
+import jakarta.persistence.Entity;
+import jakarta.persistence.OneToMany;
 
 
 @Entity
