@@ -2,7 +2,7 @@ package com.doiscs.sisman.domain.model.entity;
 
 import java.time.LocalDate;
 
-import javax.persistence.Entity;
+import jakarta.persistence.Entity;
 
 @Entity
 public class SamplePerson extends AbstractEntity {
