@@ -6,13 +6,14 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
 import com.doiscs.sisman.domain.model.entity.Equipamento;
+import org.springframework.stereotype.Repository;
 
+@Repository
 public interface EquipamentoRepository extends JpaRepository<Equipamento, Integer> {
 
-    @Query("SELECT e FROM Equipamento e WHERE e.nome LIKE COMCAT ('%', :nome, '%') ORDER BY nome")
     List<Equipamento> findByNome(String nome);
 
-    @Query("SELECT e FROM Equipamento e ORDER BY nome")
+    @Query("SELECT e FROM Equipamento e ORDER BY e.nome")
     List<Equipamento> findByAll();
 
 }
